Add a working image
The previous image didn't link to anything in particular. Linked to a functional image from the legislature's website. (Just discovered the OpenStates data while working on a bill-tracking site for my organization, ProgressNow CO. I'm not great at coding, but know enough to poke around. I may edit more CO legislators with information and try and learn how to connect to the API.)